Move ImageToLatentsInvocation to its own file. No functional changes.
